modular structure. go async
