+++ front-end/manifest.json
@@ -21,7 +21,7 @@
   "content_scripts": [
     {
       "matches": ["*://*.youtube.com/*"],
-      "js": ["content-scripts/content.js"],
+      "js": ["content-scripts/requests.js", "content-scripts/content.js"],
       "css": ["content-scripts/content.css"]
     }
   ],
