+++ front-end/popup/index.html
@@ -122,9 +122,6 @@
           <div class="header__logo">
               <span class="header__logo-title">Welcome</span>
           </div>
-          <div class="header__back">
-              <span class="back-link"><img id="back-arrow" src="../assets/img/icons/back.png"></span>
-          </div>
       </header>
       <main class="main-login">
           <section class="main__section_find">
@@ -134,7 +131,7 @@
       </main>
   </div>
 
-    <script src="popup.js"></script>
+    <script type="module" src="popup.js"></script>
 
 </body>
 </html>
